# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Learning Loop Agent API
  version: 0.2.0
  description: |
    AI 学习工具的目标 HTTP 契约。它覆盖 Ticket 01-16，Ticket 17-21 复用其中定义的工作流契约，
    是后端实现和前端集成的共同输入，
    不表示当前后端已经完成。除文件上传与下载外，请求和响应均使用 application/json。
    v0.2 使用 chat_style 表示无阶段门禁的对话风格，并用 api_format 显式选择模型调用协议。
    旧 learning_mode、SocraticState 和阶段型 intent 只用于迁移期兼容。
  x-contract-status: target
servers:
  - url: http://127.0.0.1:4173
    description: 本地开发服务
security: []
...
paths:
...
  /api/subjects/{subject_id}/exam-blueprints:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: listExamBlueprints
      tags: [Blueprints]
      summary: 列出组卷蓝图
      x-tickets: ['08']
      responses:
        '200':
          description: 组卷蓝图列表
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamBlueprintList}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      operationId: parseExamBlueprint
      tags: [Blueprints, Operations]
      summary: 把自然语言要求解析为组卷蓝图
      description: 只创建蓝图，不生成题目。
      x-tickets: ['08']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/ExamBlueprintPromptInput}
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/exam-blueprints/{blueprint_id}:
    parameters:
      - $ref: '#/components/parameters/BlueprintId'
    get:
      operationId: getExamBlueprint
      tags: [Blueprints]
      summary: 读取组卷蓝图
      x-tickets: ['08']
      responses:
        '200':
          description: 组卷蓝图
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamBlueprint}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: updateExamBlueprint
      tags: [Blueprints]
      summary: 修改组卷蓝图
      description: 仅修改标题时保持当前状态；修改题型、分值或资料范围会把已确认蓝图恢复为 draft，必须重新确认。
      x-tickets: ['08']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/ExamBlueprintPatch}
      responses:
        '200':
          description: 已修改并重新校验约束
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamBlueprint}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    delete:
      operationId: deleteExamBlueprint
      tags: [Blueprints]
      summary: 删除未被试卷草稿引用的组卷蓝图
      x-tickets: ['08']
      responses:
        '204': {description: 已删除}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/exam-blueprints/{blueprint_id}/confirm:
    parameters:
      - $ref: '#/components/parameters/BlueprintId'
    post:
      operationId: confirmExamBlueprint
      tags: [Blueprints]
      summary: 确认组卷蓝图
      description: 有 error 级约束冲突时返回 409，不触发题目生成。
      x-tickets: ['08']
      responses:
        '200':
          description: 已确认
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamBlueprint}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/exam-blueprints/{blueprint_id}/generate:
    parameters:
      - $ref: '#/components/parameters/BlueprintId'
    post:
      operationId: generateExamDraft
      tags: [Blueprints, ExamDrafts, Operations]
      summary: 从已确认蓝图开始增量组卷
      description: 立即创建带题目槽位的试卷草稿，再逐题更新状态。
      x-tickets: ['09']
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/subjects/{subject_id}/exam-drafts:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: listExamDrafts
      tags: [ExamDrafts]
      summary: 列出试卷草稿
      x-tickets: ['09']
      responses:
        '200':
          description: 试卷草稿列表
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamDraftList}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/exam-drafts/{draft_id}:
    parameters:
      - $ref: '#/components/parameters/DraftId'
    get:
      operationId: getExamDraft
      tags: [ExamDrafts]
      summary: 读取试卷草稿和逐题状态
      x-tickets: ['09']
      responses:
        '200':
          description: 试卷草稿
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamDraft}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: updateExamDraft
      tags: [ExamDrafts]
      summary: 修改试卷标题、说明或题目顺序
      x-tickets: ['09']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/ExamDraftPatch}
      responses:
        '200':
          description: 已修改
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamDraft}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    delete:
      operationId: deleteExamDraft
      tags: [ExamDrafts]
      summary: 删除试卷草稿
      x-tickets: ['09']
      responses:
        '204': {description: 已删除}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/exam-drafts/{draft_id}/questions/{question_id}:
    parameters:
      - $ref: '#/components/parameters/DraftId'
      - $ref: '#/components/parameters/QuestionId'
    put:
      operationId: replaceDraftQuestion
      tags: [ExamDrafts]
      summary: 人工替换一道题
      x-tickets: ['09']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/QuestionInput}
      responses:
        '200':
          description: 已替换并重新校验题目依据
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/DraftQuestion}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    delete:
      operationId: deleteDraftQuestion
      tags: [ExamDrafts]
      summary: 删除一道题
      x-tickets: ['09']
      responses:
        '204': {description: 已删除，不影响其他题目}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/exam-drafts/{draft_id}/questions/{question_id}/retry:
    parameters:
      - $ref: '#/components/parameters/DraftId'
      - $ref: '#/components/parameters/QuestionId'
    post:
      operationId: retryDraftQuestion
      tags: [ExamDrafts, Operations]
      summary: 单独重试一道题
      x-tickets: ['09']
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/exam-drafts/{draft_id}/publish:
    parameters:
      - $ref: '#/components/parameters/DraftId'
    post:
      operationId: publishExamDraft
      tags: [ExamDrafts, Exams]
      summary: 确认并保存为试卷
      description: 含 failed 或 needs-review 题目时必须显式确认；不会静默保存为可靠题目。
      x-tickets: ['09']
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/PublishExamInput}
      responses:
        '201':
          description: 已保存试卷
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Exam}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/exam-drafts/{draft_id}/revision-proposals:
    parameters:
      - $ref: '#/components/parameters/DraftId'
    get:
      operationId: listDraftRevisionProposals
      tags: [ExamDrafts]
      summary: 列出草稿修改提案
      x-tickets: ['13']
      responses:
        '200':
          description: 修改提案列表
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/DraftRevisionProposalList}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      operationId: createDraftRevisionProposal
      tags: [ExamDrafts, Operations]
      summary: 让 AI 生成草稿修改预览
      description: 只生成差异提案，不直接修改草稿。
      x-tickets: ['13']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/DraftRevisionProposalInput}
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/draft-revision-proposals/{proposal_id}/apply:
    parameters:
      - $ref: '#/components/parameters/ProposalId'
    post:
      operationId: applyDraftRevisionProposal
      tags: [ExamDrafts]
      summary: 应用草稿修改提案
      x-tickets: ['13']
      responses:
        '200':
          description: 已应用到草稿
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamDraft}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/draft-revision-proposals/{proposal_id}/discard:
    parameters:
      - $ref: '#/components/parameters/ProposalId'
    post:
      operationId: discardDraftRevisionProposal
      tags: [ExamDrafts]
      summary: 放弃草稿修改提案
      x-tickets: ['13']
      responses:
        '204': {description: 已放弃，草稿保持不变}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/subjects/{subject_id}/exams:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: listExams
      tags: [Exams]
      summary: 列出试卷
      x-tickets: ['09', '10']
      responses:
        '200':
          description: 试卷列表
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamList}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/subjects/{subject_id}/missed-questions:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: listMissedQuestions
      tags: [Attempts]
      summary: 列出科目内跨试卷错题投影
      description: 只读投影，按考点分组。仅统计带反馈的作答；同一题多次答错只保留最近一次。不含答案、解析或反馈明细。
      x-tickets: ['30']
      responses:
        '200':
          description: 按考点分组的错题列表
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/MissedQuestionList}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/exams/{exam_id}:
    parameters:
      - $ref: '#/components/parameters/ExamId'
    get:
      operationId: getExam
      tags: [Exams]
      summary: 读取用于编辑的完整试卷
      description: 包含答案和解析；答题界面必须使用 Attempt.paper，不能用本接口构造考试模式页面。
      x-tickets: ['09', '13']
      responses:
        '200':
          description: 完整试卷
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Exam}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      operationId: replaceExamDocument
      tags: [Exams]
      summary: 保存人工编辑后的结构化试卷
      description: base_version_id 必须等于当前版本；成功后创建新版本。
      x-tickets: ['13']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/ExamDocumentReplaceInput}
      responses:
        '200':
          description: 已创建新版本
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Exam}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    patch:
      operationId: updateExam
      tags: [Exams]
      summary: 重命名试卷
      description: 仅修改当前试卷标题，不创建新版本，也不进入撤销栈。
      x-tickets: ['13']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/ExamPatch}
      responses:
        '200':
          description: 已修改标题
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Exam}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    delete:
      operationId: deleteExam
      tags: [Exams]
      summary: 删除试卷
      x-tickets: ['09']
      responses:
        '204': {description: 已删除；版本、修改提案、作答记录和导出任务一并删除}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/exams/{exam_id}/versions:
    parameters:
      - $ref: '#/components/parameters/ExamId'
    get:
      operationId: listExamVersions
      tags: [Exams]
      summary: 列出试卷版本
      x-tickets: ['13']
      responses:
        '200':
          description: 版本列表
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamVersionList}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/exams/{exam_id}/versions/{version_id}/restore:
    parameters:
      - $ref: '#/components/parameters/ExamId'
      - $ref: '#/components/parameters/ExamVersionId'
    post:
      operationId: restoreExamVersion
      tags: [Exams]
      summary: 从旧版本恢复为一个新版本
      x-tickets: ['13']
      responses:
        '200':
          description: 已恢复
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Exam}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/exams/{exam_id}/undo:
    parameters:
      - $ref: '#/components/parameters/ExamId'
    post:
      operationId: undoExamChange
      tags: [Exams]
      summary: 撤销最近一次试卷修改
      x-tickets: ['13']
      responses:
        '200':
          description: 已撤销
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Exam}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/exams/{exam_id}/redo:
    parameters:
      - $ref: '#/components/parameters/ExamId'
    post:
      operationId: redoExamChange
      tags: [Exams]
      summary: 重做最近一次撤销
      x-tickets: ['13']
      responses:
        '200':
          description: 已重做
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Exam}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/revision-proposals/{proposal_id}:
    parameters:
      - $ref: '#/components/parameters/ProposalId'
    get:
      operationId: getExamRevisionProposal
      tags: [Exams]
      summary: 读取修改提案和差异预览
      x-tickets: ['13']
      responses:
        '200':
          description: 修改提案
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ExamRevisionProposal}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  parameters:
    SubjectId: {name: subject_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
...
    BlueprintId: {name: blueprint_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
    DraftId: {name: draft_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
    QuestionId: {name: question_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
    ExamId: {name: exam_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
...
    ProposalId: {name: proposal_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
...
  responses:
    OperationAccepted:
      description: 已接受异步任务
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/OperationAccepted}
    NotFound:
      description: 资源不存在或不属于当前科目空间
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
    Conflict:
      description: 当前资源状态不允许此操作
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
    ValidationFailed:
      description: 请求字段或领域约束校验失败
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}